refactor: 폴더 정리 및 WeatherDataManager 타입 추가
