+++ Diary/Common/CoreData/ContentsEntity+CoreDataClass.swift
