+++ Diary/Common/CoreData/ContentsEntity+CoreDataProperties.swift
