+++ Diary/Common/CoreData/CoreDataManager.swift
