+++ Diary/Common/Error/DiaryError.swift
