+++ Diary/Common/Error/NetworkError.swift
