+++ Diary/Common/Extension/NotificationName+.swift
