+++ Diary/Common/Model/WeatherDTO.swift
