+++ Diary/Common/Utility/LocationManager.swift
