+++ Diary/Presentation/DiaryList/Protocol/IdentifierType.swift
